requirements.txt
aiogram==3.4.1
PyMySQL==1.1.0
python-dotenv==1.0.1
